Adds scroll-aware landing header
- Extracts the landing page header into a dedicated client component.
- Hides the fixed header while scrolling down and restores it on upward scroll or near the top, reducing visual clutter while keeping key actions accessible.
